requirements.txt
pyyaml>=6.0.1,<7.0.0
pybit==5.7.0
websocket-client==1.7.0
pydantic==2.7.0
pytest==8.2.0

